requirements.txt
osmnx>=1.9
networkx>=3.3
ortools>=9.10
folium>=0.16
streamlit>=1.35
streamlit-folium>=0.20
elevation
rasterio>=1.3
scipy>=1.13
groq>=0.9
requests>=2.31
geopandas>=0.14
shapely>=2.0
numpy>=1.26
pandas>=2.2
python-dotenv>=1.0
Pillow>=10.0
